Three boxes
Chef has 33 boxes of sizes AA, BB, and CC respectively. He puts the boxes in bags of size DD ( A≤B≤C≤DA≤B≤C≤D ). Find the minimum number of bags Chef needs so that he can put each box in a bag. A bag can contain more than one box if the sum of sizes of boxes in the bag does not exceed the size of the bag.

Input Format
The first line contains TT denoting the number of test cases. Then the test cases follow.
Each test case contains four integers AA, BB, CC, and DD on a single line denoting the sizes of the boxes and bags.

Output Format
For each test case, output on a single line the minimum number of bags Chef needs.