Create minimum.c
program in c given an array of integers where each element represents the max number of steps that can be made forward from that element. Write a function to return the minimum number of jumps to reach the end of the array (starting from the first element). If an element is 0, it cannot move through that element. If the end isn’t reachable, return -1.